Make game framerate-independent with delta-time scaling
Replace fixed 12fps RegulateSpeed timing with delta-time infrastructure.
All movement, animations, and timers now scale correctly at any framerate.

Systems changed:
- Player movement: ScaleByDT applied in ResultsHandler after ice/swamp
- Seeker movement: temporary speed scaling in AnimateSeekers loop
- Bullet movement: scaled distance per frame in MoveABullet
- Animation rates: ScaleByDT on anim_frame_rate in AdvanceFrame
- Angel fly-away: UpdateDeltaTime replaces RegulateSpeed
- Ice physics: float ice_delay counters with dt-scaled increments
- Frame timers (supress_repeats, bogged_down_in_swamp, InputGenerator
  duration): converted to millisecond-based with FramesToMillis/TickDownTimer
- Seeker stranded/immobile counters: float with dt-scaled increments

+++ src/seeker.h
@@ -297,8 +297,8 @@ typedef struct dude
 	int32				alternate_hazard_direction;
 	int32				move_status;
 	int32				move_counter;
-	int32				stranded_counter;
-	int32				immobile_counter;
+	float				stranded_counter;
+	float				immobile_counter;
 	int32				reversal_counter;
 	int32				previous_x;
 	int32				previous_y;
